Made the replace_cleanup function more robust
Now, rather than removing the error population and calling the learning
function from the nengo learning template, simply removes the cleanup
population and connects it up to the same terminations/origins that the
previous one was connected to. Also, imposing a restriction that the
dimensions and the number of neurons of the new cleanup have to be the
same as that of the previous cleanup. We don't really need it for other
cases, and these are much easier to support.